Fix: workflow_run pull_requests id/number types
The webhook payload for the workflow_run event types (completed,
in-progress & requested) each define the id and number of pull requests
as `number` where they clearly should be `integer` values.

+++ githubkit/versions/ghec_v2022_11_28/models/group_0837.py
@@ -402,8 +402,8 @@ class WebhookWorkflowRunInProgressPropWorkflowRunPropPullRequestsItems(GitHubMod
     head: WebhookWorkflowRunInProgressPropWorkflowRunPropPullRequestsItemsPropHead = (
         Field()
     )
-    id: float = Field()
-    number: float = Field()
+    id: int = Field()
+    number: int = Field()
     url: str = Field()
